+++ src/app/fdctl/Local.mk
@@ -80,6 +80,7 @@ $(call add-objs,configure/hyperthreads,fd_fdctl)
 $(call add-objs,configure/ethtool-channels,fd_fdctl)
 $(call add-objs,configure/ethtool-gro,fd_fdctl)
 $(call add-objs,configure/ethtool-loopback,fd_fdctl)
+$(call add-objs,configure/sysfs-busypoll,fd_fdctl)
 
 ifdef FD_HAS_NO_AGAVE
 ifdef FD_HAS_SECP256K1
