+++ src/app/fdctl/configure/configure.h
@@ -69,6 +69,7 @@ extern configure_stage_t fd_cfg_stage_hyperthreads;
 extern configure_stage_t fd_cfg_stage_ethtool_channels;
 extern configure_stage_t fd_cfg_stage_ethtool_gro;
 extern configure_stage_t fd_cfg_stage_ethtool_loopback;
+extern configure_stage_t fd_cfg_stage_sysfs_busypoll;
 
 extern configure_stage_t * STAGES[];
