+++ src/disco/net/Local.mk
@@ -1,6 +1,6 @@
 ifdef FD_HAS_SSE
 ifdef FD_HAS_ALLOCA
 $(call add-hdrs,fd_net_tile.h)
-$(call add-objs,fd_net_tile fd_net_tile_topo,fd_disco)
+$(call add-objs,fd_net_tile_topo,fd_disco)
 endif
 endif
