+++ src/disco/net/xdp/Local.mk
@@ -1,0 +1,5 @@
+ifdef FD_HAS_SSE
+ifdef FD_HAS_ALLOCA
+$(call add-objs,fd_xdp_tile fd_xdp_tile_softirq fd_xdp_tile_poll,fd_disco)
+endif
+endif
